apps/api/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# Async
httpx==0.26.0
aiofiles==23.2.1

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Password hashing (argon2)
argon2-cffi==23.1.0
